Extracted history section into new component
