+++ src/components/Home.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
-  import { Accordion, AccordionItem } from "carbon-components-svelte";
   import type { Entry } from "src/types/entry.type";
 
-  let startOfToday = new Date().setHours(0,0,0,0);
+  import History from "./History.svelte";
+
   let displayedList = [];
 
   const TASKLIST = "tasklist";
@@ -58,7 +58,6 @@
   let categoryInput = '';
 
   let displayEntries = displayStoredEntries();
-  let historyResults = organizeHistoryPromise();
 
   // create new Entry object
   // if Entry object is missing info, the following functions should fail
@@ -302,21 +301,6 @@
     console.log("categoryList in displayStoredEntries", categoryList);
   }
 
-  async function organizeHistoryPromise(): Promise<{}> {
-    const historyItems = await chrome.history.search({text: "", startTime: startOfToday});
-    
-    let organizedHistory = {};
-    for (let item of historyItems) {
-      let hostname = new URL(item.url).hostname;
-        Array.isArray(organizedHistory[hostname]) ?
-          organizedHistory[hostname].push(item) :  
-          organizedHistory[hostname] = [item];
-    }
-
-    // console.log("Organized History", organizedHistory);
-    return organizedHistory;
-  }
-
   let obj = 
   {
     "google":  ["google.com", "google.com/1"],
@@ -326,9 +310,6 @@
 </script>
 
 <div>
-  <!-- <p>
-    Start of Today: <span>{new Date(startOfToday)}</span>
-  </p> -->
 
   <!-- <form on:submit|preventDefault={setEntry}>
     <input bind:value={textInput}>
@@ -400,27 +381,7 @@
   </div>
   {/if}
 
-  {#await historyResults}
-    <p>...waiting</p>
-  {:then organizedHistoryObject}
-  <Accordion size="xl" align="start">
-      {#each Object.entries(organizedHistoryObject) as [hostname, HistoryItems]}
-        <AccordionItem title="{hostname}">
-          <ul>
-            {#each Object.values(HistoryItems) as historyItem}
-              {#if historyItem.title.trim().length !== 0}
-              <li><a href="{historyItem.url}" rel="noopener noreferrer" target="_blank">
-                Item: {historyItem.title}
-              </a></li>
-              {/if}
-            {/each}
-          </ul>
-        </AccordionItem>
-      {/each}
-    </Accordion>
-    {:catch error}
-      <p style="color: red">{error.message}</p>
-  {/await}
+  <History />
 
 </div>
